+++ R/benchmark.R
@@ -70,8 +70,6 @@
 #' \donttest{
 #' library(ClassicHCAS)
 #'
-#' obs_dat <- read.csv("")
-#' prd_dat <- read.csv("")
 #'
 #'
 #' }