+++ R/calibrate.R
@@ -30,9 +30,6 @@
 #' \donttest{
 #' library(ClassicHCAS)
 #'
-#' obs_dat <- read.csv("")
-#' prd_dat <- read.csv("")
-#' histo <- read.table("histogram.txt")
 #'
 #'
 #' }