+++ NEWS.md
@@ -1,5 +1,5 @@
 # Version 0.1.2
-* Checks for availability of the `terra` package is added when inputs is a raster
+* Check for availability of the `terra` package is added when input is a raster
 
 # Version 0.1.1
 * Added the geographic distance penalty for selecting the nearest predicted neighbours (`xy_penalty` and `xy_stats` parameters)