+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 0.1.2
 * Check for availability of the `terra` package is added when input is a raster
 * improved documentation and imports
+* a bug fix in benchmarking
 
 # Version 0.1.1
 * Added the geographic distance penalty for selecting the nearest predicted neighbours (`xy_penalty` and `xy_stats` parameters)